requirements.txt
RPi.GPIO
Flask-WTF
Flask